LU-4860 lov: Handle the case of stripe size is not power 2
Calculate the end of current stripe correctly when the stripe size
is not power 2.

LU-4873 llite: avoid a deadlock in page write

For a partial page write, it will have to issue a READ RPC firstly
to get a full uptodate page. If another page is already locked by
this thread it can easily cause deadlock.

Test-Parameters: fortestonly
Signed-off-by: Jinshan Xiong <jinshan.xiong@intel.com>
Change-Id: I5ff825ea06ef4a72a39b97c6d4eab36ce3ecb540