LU-4795 obd: Fix bug in server_put_mount()
This patch fixes server_put_mount() to call server_deregister_mount()
only if this is the last reference. As a result, the functions
server_{get,put}_mount_2() are not needed anymore and are hence
removed. The  only caller of server_get_mount_2(),
lustre_find_lwp_by_index(), is changed to now call server_get_mount()
and server_put_mount().

Signed-off-by: Swapnil Pimpale <spimpale@ddn.com>
Change-Id: I37fd00bd477f946ac1bd8375da4899c2a1fe9e22