LU-4609 ofd: race between LFSCK and OST recovery
During the OST recovery, the client write RPC may create the missed
OST-object which will cause the real last_id to be greater than the
last_id value stored in the LAST_ID file temporarily. At that time,
if the LFSCK is scanning the system and find such inconsistency, it
will regard the LAST_ID files as crashed, then it will mark the OST
device as read-only and rebuild the LAST_ID files.

To avoid such race, before the write RPC to create the lost object,
it needs to update the LAST_ID file firstly, the LFSCK will handle
that as does for MDT-OST recovery, which will not cause the device
to be set as read-only and allow the OST recovery and LFSCK routine
check in parallel.

Signed-off-by: Fan Yong <fan.yong@intel.com>
Change-Id: I963f3c1b70c7cad0c943f2485417d2e783768bf1