LU-4895 lfsck: not create object to repair dangling by default
If the OST-object which is referenced by some MDT-object does not
exist, the layout LFSCK should not create the lost OST-object by
default; otherwise related application(s) may miss to be aware of
the bad file.

The LFSCK will allow the administrator to specify how to handle
the dangling reference case via a new option "-c" when trigger
the LFSCK:

-c off: report the inconsistency via log, but keep the dangling
        reference there without repairing. (by default)
-c on: create the lost OST-object.

Other code cleanup.

Signed-off-by: Fan Yong <fan.yong@intel.com>
Change-Id: I8fa1e5bbea63c7368e04052c88311a9ea0c6f969