LU-4675 lfsck: new magic for partially repaired file
When the layout LFSCK repairs orphan OST-object, if the parent
MDT-object was lost, then it will re-create the MDT-object and
regenerate the LOV EA and fill the target LOV EA slot with the
orphan information, and fill other slots with zero (dummy slot);
if related LOV EA slot is invalid or dummy, then it will refill
the target LOV EA slot; if the target slot exceeds current LOV
EA tail, then extend the LOV EA, and fill the gaps as dummy.

Some of the dummy LOV EA slots may cannot be re-filled finally becuase
of lost some OST-objects. And even if they can be re-filled, but there
is still some race accessing from client before the re-filling. If the
client accessed the LOV EA with dummy slots, it may cause some strange
behaviour, and may trigger LBUG() or LASSERT() on the client.

So we need to make the client to be aware of the LOV EA is incomplete.
We introduce a new LOV EA magic: LOV_MAGIC_PARTIAL for that: any time
when the LFSCK repairs the LOV EA with dummy slot(s), the LOV EA will
be marked as LOV_MAGIC_PARTIAL; when all the dummy slots are refilled,
the LOV EA will be marked as LOV_MAGIC_V1.

For a new client, it recongizes the new magic LOV_MAGIC_PARTIAL, then
it permits/forbids some opertions on the file which has dummy slots:

1) read-only opertions are permitted, such as stat/ls -l/read, and so
   on. The file's size will be the sum on the known stripes. It gives
   the administrator chance to know how much data has been recovered,
   and the administrator can dump those data via normal read tools.

2) If the modification only changes MDS-side metadata, such as chmod,
   then it is permitted.

3) unlink/rm the file which has dummy LOV EA slots is permitted.

4) For other modifications, if the modification will change something
   on OST side, such as write/touch/chown, then it will be denied.

For a old client, it will not recognize the magic LOV_MAGIC_PARTIAL.
So the LOV EA with dummy slots will be dicarded with failure, but it
will not cause the client to be crashed.

Signed-off-by: Fan Yong <fan.yong@intel.com>
Change-Id: Ic4a7fd08b229752cc155606515b6c9ddd380213e