LU-2456 lnet: DLC Feature dynamic net config
This is the third patch of a set of patches that enables DLC.

This patch adds the following features to LNET.  Currently these
features are not driven by user space.
- Adding/Deleting Networks dynamically
Two new APIs were added LNetNIDynInit() and LNetNIDynFini().
LNetNIDynInit() starts up an NI and initializes the NI system
if it hasn't been initialized already.
LNetNIDynFini() removes an NI and uninitializes the NI system
if no more NIs are active (except for LO)

Signed-off-by: Amir Shehata <amir.shehata@intel.com>
Change-Id: I43c1c3303916b15772b19ea94f56fb277a9d87bb