LU-0000 proto: multi-slot last_rcvd
THIS IS A PROTO TO PLAY WITH IDEA.

basically the idea is to:
 - use additional log to store transno/xid/vbr/rc pairs
   (struct lsd_reply_data)
 - in-core bitmap is used to find free slots in this additional log
 - rep-ack is used to free the slots
 - all requests with transno schedule rep-ack

an interesting observation from dbench workload:
lustre-MDT0000: 236674/19601(e), diff: 236674/93694(f)
lustre-MDT0000: 0 slots, log size 11

236674 transactions were generated, 19601 were empty (open/close not
changing FS). 236674 rep-acks were requested, from which 93694 were
enforced by this patch. in theory, 19601 could be optimized do not
request rep-ack. so that would make (93694-19601)/(236674-93694-19601)
- 45% extra rep-acks compared to the master.

Signed-off-by: Alex Zhuravlev <alexey.zhuravlev@intel.com>
Change-Id: Ie551f72596b01bea000467a74ff917b574249e73