LU-4719 mdt: Don't try to print non-existent objects
mdt_dump_lmm() needs to make sure it's not trying to print
objects for a default striping from a directory, since there
are not objects in this case.

Return early if the D_INFO debug mask is disenabled, so it
doesn't iterate through the LOV/LMV layout at all.  That is
just a waste of effort.

This patch is back-ported from the following one:
Lustre-commit: ebfc6f4b54961e5e0762af7d2555f2dc5d1e1ec3
Lustre-change: http://review.whamcloud.com/9520

Signed-off-by: Jian Yu <jian.yu@intel.com>
Change-Id: I1adf4bf03956da8f6720548fbf51ab22cb1177c1