LU-4675 lfsck: new pattern flag for partially repaired file
When the layout LFSCK repairs orphan OST-object, if the parent
MDT-object was lost, then it will re-create the MDT-object and
regenerate the LOV EA and fill the target LOV EA slot with the
orphan information, and fill other slots with zero (LOV hole);
if related LOV EA slot is invalid or hole, then it will refill
the target LOV EA slot; if the target slot exceeds current LOV
EA tail, then extend the LOV EA, and fill the gaps as zero.

Some of the LOV EA holes may cannot be re-filled finally because
of lost some OST-objects. And even if they can be re-filled, but
there are still some possible race accessings from client before
the re-filling. If the client access the LOV EA with hole(s), it
may cause some strange behaviour, such as trigger LBUG()/LASSERT()
on the client.

So we will make the client to be aware of the LOV EA is incomplete.
We introduce a new LOV EA pattern flag LOV_PATTERN_F_HOLE for that:
any time when the LFSCK repairs the LOV EA with hole(s), the LOV EA
will be marked as LOV_PATTERN_F_HOLE; when all the holes in the LOV
EA are refilled, the LOV_PATTERN_F_HOLE will be dropped.

For a new client, it recongizes the pattern flag LOV_PATTERN_F_HOLE,
then it can permit/forbid some opertions on the file with LOV holes:

1) getattr/getxattr opertions are permitted, such as stat/ls -l, and
   so on. The file size is the sum on the known stripes. So it gives
   the administrator chance to know how much data has been recovered.

2) Normal read the file with LOV EA hole is permitted, but will get
   EIO failure when try to read data from the dummy object which is
   corresonding to the LOV EA hole. So the users can dump recovered
   data via some common read tools, such as "dd conv=sync,noerror".

3) If the modification only changes MDS-side metadata, such as chmod,
   then it is permitted.

4) unlink/rm the file which has LOV EA holes is permitted.

5) For other modifications, if the modification will change something
   on OST side, such as write/touch/chown, then it will be denied.

For a old client, since it will not recognize the new pattern flag
LOV_PATTERN_F_HOLE. So the LOV EA with holes will be dicarded with
failure, but it will not cause the client to be crashed.

Signed-off-by: Fan Yong <fan.yong@intel.com>
Change-Id: Ic4a7fd08b229752cc155606515b6c9ddd380213e