LU-4557 ext4: recalc percpu counters after journal
the percpu counters is initialized before replaying the journal,
but after the journal it should be recalculate to deal with the
possibility of the percpu counters getting updated by the journal
replay.

the issue has been fixed in the new kernels (RHEL6 2.6.32-431.5.1,
SLES11SP3 3.0.101-0.15), then it is not needed for these kernels.

Lustre-commit: ecaaa1bd05ab13feb4794f4f4c1f6bc64da7e8ad
Lustre-change: http://review.whamcloud.com/9908

Signed-off-by: Hongchao Zhang <hongchao.zhang@intel.com>
Signed-off-by: Shuichi Ihara <sihara@ddn.com>
Change-Id: I2215efd74af3ba569ec08ae5b0454ca9f4a49673