LU-5301 tests: increase timeout for sanity-lfsck tests
Increase the time waited for sanity-lfsck tests for LFSCK to complete
or perform some action.  When running in VMs or with different
filesystems, the time taken by LFSCK stages can vary noticeably.
Use wait_timeout_facet() so that the wait is as short as possible.

Consolidate repeated commands to get the LFSCK status for ease of
reading the tests.

Update wait_update() to follow proper coding style (lower-case local
variables, avoid name clashes with bash builtins), and add a new
comment block for it.  Add argument handling to wait_update_facet().

Remove redundant calls to clear fail_loc at the end of a test, since
this is done in run_one() after every test.

Signed-off-by: Andreas Dilger <andreas.dilger@intel.com>
Change-Id: Ia42ba28ba5e60f7c127fa930d9025244a5500c1e