LU-4840 lfs: Use file lease to implement migration.
Rely on exclusive open instead of group lock to migrate a file.
Implemented exclusive close operation to atomically put a lease, swap
two layouts and close a file. This allows race-free migrations
implemented with leases instead of grouplock.

Deprecate the --block switch of lfs migrate, and make the caller
responsible for retrying on failure (EBUSY, EAGAIN).

Signed-off-by: Henri Doreau <henri.doreau@cea.fr>
Change-Id: Ie420e1998cae03928ae24834070c16642e8cd3b9