LU-4931 ladvise: Add feature of giving file access advices
This patch adds a new feature which provides new APIs and utils
to give advices about the access pattern of Lustre file with the
purpose of performance improvement. It can pass advices directly
through Lustre client to server/storage side. Currently, the only
supported advice is to prefecth data to memory cache of OSS.

Signed-off-by: Li Xi <lixi@ddn.com>
Change-Id: I6de14ed87199949963360d393897ff79826c9819