LU-4932 doc: desgin doc for LFSCK
LFSCK design documentation has previously been recorded in an wiki
external to the souce code. This commit adds a first draft of the design
features of the LFSCK feature. A new directory has been created to house
this and other future design documents modelled after the Linux Kernel
'Documentation' directory.

Change-Id: I598261efd7fc4ac71a32a301b8afe8c60b592ace
Signed-off-by: Richard Henwood <richard.henwood@intel.com>