LU-4713 tests: disable sanity test_237
For now this test does nothing in Centos/RHEL due to the lack of
the feature in the kernel and always fails in SLES kernels where
the kernel feature exists but glibc support is missing.
Since that is the case it should be disabled as it's adding
no value or useful information.

Test-Parameters: clientdistro=sles11sp3 testlist=sanity
Signed-off-by: Bob Glossman <bob.glossman@intel.com>
Change-Id: I09c4ff521a797f8054511a78cca86cbd3f1379f9