LU-4943 obdclass: detach MGC dev on error
lustre_start_mgc() creates MGC device, if error happens later on
ll_fill_super(), this device is still attached, and later mount
fails by keep complaining that the MGC device's already in the
client node.

It turns out that the device was referenced by mgc config llog data
which is arranged in the mgc lock requeue threadn for re-trying to
get mgc lock, and in normal case, this llog reference only released
in mgc_blocking_ast() when the system is umount.

This patch handles the error path when mgc connect fails, it releases
the config llog data reference in config_log_end() since there will
be no mgc_blocking_ast() happen in this case.

Signed-off-by: Bobi Jam <bobijam.xu@intel.com>
Change-Id: I3cd7afbdf733fd522aed4157098fbe1c72332a0a