LU-4969 nodemap: nodemap key hash comparison
Changed the keymap comparison for the nodemap hash to the fix that
Andreas suggested, and fixed a bug with the idmap tree that resulted
in an incorrect deletion of idmaps.

Updated the sanity-sec.sh tests to allow for nodemaps to be tested
in amounts larger than the initial hash size to check the fix.

Signed-off-by: Joshua Walgenbach <jjw@iu.edu>
Change-Id: I464fa8c97cf77b57e7a0918a2bb5aee935bfc891