LU-4606 utils: Split zfs functionality into module for mount
Pull the ZFS functionality of mount/mkfs out into a loadable module so
that it can be dynamically linked into mount (instead of dlopen'ing a
hardcoded libzfs version).  This allows libzfs version to change
without altering lustre source.  This has the benefit of making
lustre-osd-zfs explicitly requires libzfs (which it needs to do
mount/mkfs operations).

Adjust build system to allow building of loadable module.

This functionality should be easily extensible to allow splitting
ldiskfs out into it's own loadable module (or other modules to be
build in the future).

Signed-off-by: Nathaniel Clark <nathaniel.l.clark@intel.com>
Change-Id: I8fc0f41161dc08405c1f951f12e675f364ff62c3