LU-9 qos: skeleton for weighted round-robin stripe allocator
This patch is a rebase of Isami Romanowski's original patch.

It includes the beginnings of the weighted round-robin
allocator, which is intended to replace the current two-phase stripe
allocation (with disjoint round-robin and QoS steps).

Currently, however, the behavior of the weighted round-robin allocator
is effectively identical to that of the original round-robin
allocator. There are some steps needed for the new allocator to start
showing improved behavior:

    * Per-pool average weights need to be computed. There is code
      already which exists to do this, but adding in the op_avg_weight
      field to struct ost_pool caused bizarre test failures.
      As of now, the weight threshold during allocation is always set
      to 0, so the weight of an OST is irrelevant. When these averages
      are stored, the threshold during allocation should be set to the
      average OST weight in the pool.
    * Weight computations are currently done within the allocator code
      path, which is not necessarily desirable. Better behavior would
      be to compute weight for an OST at every statfs call.

The Bugzilla thread about this feature, which can be accessed from the
JIRA page for the LU-9 ticket, has more information about some of the
desired behavior of the new allocator.

Signed-off-by: Swapnil Pimpale <spimpale@ddn.com>
Signed-off-by: Isami Romanowski <isami@whamcloud.com>
Change-Id: I0b4395f4f98934d9a05b66c07cba205788e1be92