LU-2096 lprocfs: symlinks for forward compatibility
Currently, the "ofd" device is still named "obdfilter" in the Lustre
configuration and in /proc/fs/lustre.  It would be better to name it
"ost" which is has been the long term goal but retain the "ofd"
module name.

The per-OST device statistics should really be under "ost/{target}",
like the per-MDT statistics are under "mdt/{target}" already in 2.x.

Similarly, the "OSS" statistics have always been under "ost/OSS", but
would be better in their own "oss" subdirectory much like what is done
for "MDS" statistics.

Add symlinks in lprocfs for "ost/OSS -> oss", "obdfilter -> ofd -> ost",
and "obdfilter/{target} -> ost/{target}" for forward compatibility so
that when we finally are able to make these changes to the module names
userspace tools and scripts can have a smoother transition between
Lustre versions.

Signed-off-by: Andreas Dilger <andreas.dilger@intel.com>
Signed-off-by: John L. Hammond <john.hammond@intel.com>
Signed-off-by: James Simmons <uja.ornl@gmail.com>
Change-Id: Ia0835e39aba7c8a540a6247bd3d56bc14e500c1e