LU-4974 lod: use common hash function for OST pool
Use the common djb2 hash function for OST pool lookup.  This is used
for string-based hashing in other parts of the code.  It doesn't make
sense to have a custom hash function just for OST pool lookup, since
there aren't expected to be a huge number of pools and there is no
expectation that one of these functions is better than the other.

Signed-off-by: Andreas Dilger <andreas.dilger@intel.com>
Change-Id: I501a42bc169243204dfc847c0388240fc2500c1e