LU-5026 lprocfs: Add lbug_on_eviction option
Add an lbug_on_eviction proc interface.  When it is set to a non-zero
value on a client, it will cause the client to LBUG whenever it is
evicted by the server. Note, an MDS is a client to OSTs, and every
server is a client of MGS. Thus, it is probably desireable to leave
this set to zero on servers.

Signed-off-by: Ryan Haasken <haasken@cray.com>
Change-Id: I2f694e4bc1b1ec9daf52f8a6ac46ad87f40f810f