LU-5026 lprocfs: Add lbug_on_eviction option
Add an lbug_on_eviction proc interface.  When it is set to a non-zero
value, it will cause a node to LBUG on eviction.  On the client, this
means whenever it is evicted by a server.  On the server, this means
whenever it evicts a client due to expired locks.  The server will not
LBUG when it evicts a client for ping timeouts or when a user requests
client eviction through the evict_client proc interface.

Signed-off-by: Ryan Haasken <haasken@cray.com>
Change-Id: I2f694e4bc1b1ec9daf52f8a6ac46ad87f40f810f