LU-5018 libcfs: don't depend on configure for lustre_user.h
Don't depend on configure to detect the presence of <linux/types.h>
in order to define the __u8, __u16, __u32, and __u64 types.  Instead,
use __linux__ to decide that this header is present, which it should
be for all Linux distros if the devel packages are installed, which
would be a prerequisite for building on a node in the first place.

The inclusion of <linux/types.h> fixes the e2fsprogs build for lfsck,
since it checks for the _LINUX_TYPES_H define and avoids defining its
own __u* types in that case.

Signed-off-by: Andreas Dilger <andreas.dilger@intel.com>
Change-Id: Ida0915fe1b5829f3378e459c782844cf9df060e6