LU-2059 mgs: don't fail on missing params log
The mgc_process_cfg_log() fails if 'params' log is not
accessinble and has no local copy. In fact that log is
optional and may be just empty.

Patch allows 'params' log to be empty and checks also for
'params' log in other places to avoid wrong error messages.
There was also deprecated code in mgs_write_log_direct_all()
which creates lustre-params log which is not used by anyone now.
Third change is removal of exceptions in conf-sanity.sh and
insanity.sh related to LU-2059

The only exception is insanity test_1 which is still failed due
to other reason.

Signed-off-by: Mikhail Pershin <mike.pershin@intel.com>
Change-Id: I1ad024a87b0a10665f9669fa8ff7cca2d07a09fe