LU-2059 mgs: don't fail on missing params log
The mgc_process_cfg_log() fails if 'params' log is not
accessinble and has no local copy. In fact that log is
optional and may be just empty.

Patch allows 'params' log to be empty and checks also for
'params' log in other places to avoid wrong error messages.
There was also deprecated code in mgs_write_log_direct_all()
which creates lustre-params log which is not used by anyone now.
Third change is removal of exceptions in conf-sanity.sh and
insanity.sh related to LU-2059

The only exception is insanity test_1 which is still failed due
to other reason.

Lustre-change: http://review.whamcloud.com/10311
Lustre-commit: 864fc9daac267819f5e3bdebef6cdac4c6325626

With follow-on patch:

Lustre-change: http://review.whamcloud.com/10589
Lustre-commit: 8c21986e79f50131b0f381e5fe0311294328d660

Signed-off-by: Mikhail Pershin <mike.pershin@intel.com>
Reviewed-by: Andreas Dilger <andreas.dilger@intel.com>
Reviewed-by: Jian Yu <jian.yu@intel.com>
Reviewed-by: Fan Yong <fan.yong@intel.com>
Change-Id: I84d1419edca2e4a313721a5213c128b710943c9d
Signed-off-by: Nathaniel Clark <nathaniel.l.clark@intel.com>