LU-4367 close: combine close in unlink RPC
With open lock taken, close will cache file handle instead of
release it on MDS, but if unlink is followed by close, a real
close will be triggered before unlink, which will hurt unlink
performance.

This patch combines close into unlink RPC when SOM is disabled,
because current SOM implementation has flaws, and supporting it
will cause unexpected complexity. In the future if SOM is mainly
implemented on server side, this feature should coexist with it.

* client packs file handle in unlink reint record.
* MDT close file handle before unlinking.

Signed-off-by: Lai Siyao <lai.siyao@intel.com>
Change-Id: I7f4b3f5eae30183ac269ac50b8f196c7ab73c76d